% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesCoxMcmc.R
\name{bayesCoxMcmc}
\alias{bayesCoxMcmc}
\title{Get the MCMC Samples from \code{bayesCox}}
\usage{
bayesCoxMcmc(object, parts = c("h0", "coef"), ...)
}
\arguments{
\item{object}{A \code{bayesCox} object}

\item{parts}{A character vector specifying the parts to be exacted from the
MCMC output text file produced by \code{bayesCox}.  One or more
following options can be specified: \code{"h0"} for baseline hazard
function, \code{"coef"} for covariate coefficients, \code{"nu"} for
sampled latent variance of coefficients, \code{"jump"} for indicators of
jumps, and \code{"all"} for all of the above.  The default value is
\code{c("h0", "beta")}.}

\item{...}{Other arguments that are not used now.}
}
\description{
Returns the MCMC samples produced by
\code{bayesCox} into data frames.
}
