% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.bayesCox}
\alias{is.bayesCox}
\title{Is \code{bayesCox} Class Object}
\usage{
is.bayesCox(object)
}
\arguments{
\item{object}{An R object.}
}
\description{
Checks if the input object is of Class \code{bayesCox}.
}
