% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump.R
\name{jump}
\alias{jump}
\alias{jump.bayesCox}
\title{Generic function for jump information}
\usage{
jump(object, ...)

\method{jump}{bayesCox}(object, ...)
}
\arguments{
\item{object}{An object returned by function \code{bayesCox}.}

\item{...}{Other arguments.}
}
\value{
A data.frame with 3 columns \code{("Count", "Iter", "Cov")}, where
\code{"Count"} is the number of coefficient pieces (jumps) for each
iteration; \code{Iter} is the iteration number; \code{Cov} contains the
character values of the covariates.
}
\description{
Generic function for jump information
}
\section{Methods (by class)}{
\itemize{
\item \code{bayesCox}: Extract Jump Information from Bayesian Dyanmic Model

Extract number of coefficient pieces from \code{bayesCox} fitting results,
and summarize them into a data frame. It is only applicable when
\code{model="Dynamic"} is specified.
}}

\examples{
## See the examples in bayesCox

}
\seealso{
\code{\link{bayesCox}}, \code{\link{plotJumpTrace}}, and
\code{\link{plotJumpHist}}.
}
\keyword{bayesCox}
\keyword{extract}
\keyword{jump}
