% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherExports.R
\name{other-exports}
\alias{other-exports}
\alias{Surv}
\title{Exported Functions from Other Packages}
\description{
This page is a placeholder for function(s) exported by \code{dynsurv} but
internally imported from other packages. Please refer to the original
documentation of the imported function for help.
}
\details{
\itemize{

\item \code{\link[survival]{Surv}}: imported from the \code{survival}
package.

}
}
