% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynsurv-package.R
\docType{package}
\name{dynsurv}
\alias{dynsurv}
\alias{dynsurv-package}
\title{dynsurv: Time-varying coefficient models for interval censored data.}
\description{
Functions to fit time-varying coefficient models for interval censored and
right censored survival data. Three major approaches are implemented:
\enumerate{
\item Bayesian Cox model with time-independent, time-varying or dynamic
    coefficients for right censored and interval censored data;
\item Spline based time-varying coefficient Cox model for right censored
    data;
\item Transformation model with time-varying coefficients for right censored
    data using estimating equations.
}
}
