% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotJump}
\alias{plotJump}
\alias{plotJumpTrace}
\alias{plotJumpHist}
\title{Plot Jump Information in Bayesian Dynamic Model}
\usage{
plotJumpTrace(object, ...)

plotJumpHist(object, ...)
}
\arguments{
\item{object}{A data.frame returned by function \code{jump}.}

\item{...}{Other arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotJumpTrace} plots the MCMC history of the number of pieces.
\code{plotJumpHist} plots the histogram of the number of pieces. The input
data frame is returned by function \code{jump}.
}
\examples{
## See the examples in bayesCox
}
\seealso{
\code{\link{jump.bayesCox}}.
}
\keyword{jump}
\keyword{plot}
