% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNu}
\alias{plotNu}
\title{Plot Latent Variance in Bayesian Cox Model}
\usage{
plotNu(object, ...)
}
\arguments{
\item{object}{A data.frame returned by the function \code{nu}.}

\item{...}{Other arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the latent variance \code{nu} when the hierarchical AR(1) process
prior is used for the \code{bayesCox} model. It is applicable when
\code{model="TimeVarying"} or \code{model="Dynamic"}, and
\code{coef.prior=list(type="HAR1")}. The input data frame is returned by
function \code{nu}.
}
\examples{
## See the examples in bayesCox
}
\seealso{
\code{\link{nu.bayesCox}}.
}
\keyword{latent}
\keyword{plot}
\keyword{variance}

